VAL-132: Updated logging, added Vyse, updated daily report to be more careful

+++ ValorantApp/HenrikJson/MatchPlayerJson.cs
@@ -125,7 +125,7 @@
     public class PlayerEconomyMoneyJson
     {
         public int? Overall { get; set; }
-        public int? Average { get; set; }
+        public double? Average { get; set; }
     }
     #endregion
 }
